third_party/blink/web_tests/external/wpt/fetch/private-network-access/iframe.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] private to local: failure.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] public to local: failure.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] public to private: failure.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] treat-as-public-address to local: failure.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] treat-as-public-address to private: failure.
  assert_equals: expected "timeout" but got "loaded"
Harness: the test ran to completion.
